rip out ray from beanstalkd worker.py
