+++ python/python/bystro/prs/handler.py
@@ -3,7 +3,7 @@ from typing import Any, Callable
 
 from msgspec import json
 
-from bystro.beanstalkd.worker import ProgressPublisher, get_progress_reporter
+from bystro.beanstalkd.messages import ProgressPublisher, get_progress_reporter
 from bystro.ancestry.ancestry_types import AncestryResults
 from bystro.prs.messages import PRSJobData, PRSJobResult
 from bystro.prs.preprocess_for_prs import generate_c_and_t_prs_scores
